+++ src/shinestacker/project/main_window.py
@@ -60,6 +60,7 @@ class MainWindow(ProjectHandler, QMainWindow):
         actions = {
             "&New...": self.new_project,
             "&Open...": self.open_project,
+            "Open Project As Template": self.open_template,
             "&Close": self.close_project,
             "&Save": self.save_project,
             "Save &As...": self.save_project_as,
@@ -328,6 +329,15 @@ class MainWindow(ProjectHandler, QMainWindow):
                             self.action_dialog.exec()
         elif msg != '':
             self.show_status_message(msg)
+
+    def open_template(self):
+        self.open_project()
+        jobs = self.project_jobs()
+        for job_index, job in enumerate(jobs):
+            self.selection_state.set_indices(job_index)
+            self.rename()
+            self.element_action.open_job_browse_folder_dialog()
+        self.selection_state.set_indices()
 
     def new_project(self):
         if self.check_unsaved_changes():