+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Changelog
 
-## [v1.14.x] - 2026-04-08
-**Minor fixes and updates**
+## [v1.14.5] - 2026-04-09
+**Added project template, minor fixes and updates**
+
+### Added
+- menu entry "Open Project As Template": allow users to rename jobs and customize input selection (#31)
 
 ### Fixed
 - required package version Pygments>=2.20.0 to fix a security warning
 - updated release script with more recent action versions
 
+---
 
 ## [v1.14.4] - 2026-03-22
 **Minor GUI fixes and code cleanup**